requirements.txt
yfinance
pandas
numpy
matplotlib
seaborn
pyyaml
scipy
statsmodels
dash
plotly